LayoutTests/media/modern-media-controls/forward-button/forward-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>ForwardButton</code> class.");

const forwardButton = new ForwardButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("forwardButton.element.localName", "button");
shouldBeTrue('forwardButton.element.classList.contains("icon")');
shouldBeTrue('forwardButton.element.classList.contains("forward")');
shouldBe("forwardButton.iconName", "Icons.Forward");

shouldBecomeEqual('forwardButton.element.style.webkitMaskImage.includes("macOS/forward@")', "true", finishMediaControlsTest);

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
